+++ client/model_node_pool_status_amazon.go
@@ -14,8 +14,8 @@ type NodePoolStatusAmazon struct {
 	InstanceType string `json:"instanceType,omitempty"`
 	SpotPrice string `json:"spot_price,omitempty"`
 	Autoscaling bool `json:"autoscaling,omitempty"`
-	Count float32 `json:"count,omitempty"`
-	MinCount float32 `json:"minCount,omitempty"`
-	MaxCount float32 `json:"maxCount,omitempty"`
+	Count int32 `json:"count,omitempty"`
+	MinCount int32 `json:"minCount,omitempty"`
+	MaxCount int32 `json:"maxCount,omitempty"`
 	Image string `json:"image,omitempty"`
 }
