+++ client/model_node_pools_google.go
@@ -12,7 +12,7 @@ package client
 
 type NodePoolsGoogle struct {
 	Autoscaling bool `json:"autoscaling,omitempty"`
-	Count float32 `json:"count"`
+	Count int32 `json:"count"`
 	MinCount int32 `json:"minCount,omitempty"`
 	MaxCount int32 `json:"maxCount,omitempty"`
 	InstanceType string `json:"instanceType"`
